Move wallet display and path resolution to linera-wallet
Add paths module with config_dir, wallet_path, keystore_path helpers.
Add display module with pretty_print for wallet chain details.
Remove WalletExt trait from linera-service — display logic now in
linera-wallet::display.
